msgs: move ServerNamePayload into client_hello
